fix: Allow YAML files to have no extension
The file extension wasn't used previously to determine if a file was yaml.
So the addition of JSON and thus checking of extensions introduced a
scenario where yaml files without extension resulted in an error. This
change allows resets the default behavior that a file with no extension
is assumed to be yaml.

Added integration tests to test all three variants